move validation scripts and tests to validation folder
