+++ src/networks/ai/ai_api_request.ts
@@ -153,19 +153,3 @@ export async function persistReview(
   console.log("🏁 runAiReview completed");
   return reviewWithIds;
 }
-export function removeAdditionalLineNumbersAndSymbols(
-  review: AiResponse,
-): AiResponse {
-  const sanitisedLineNumbers = review.feedback_points.flatMap((point) => {
-    const match = point.line_numbers[0].match(/\d+(?:-\d+)?/);
-    if (!match) return [];
-    return [
-      {
-        ...point,
-        line_numbers: [match[0], ...point.line_numbers.slice(1)],
-      },
-    ];
-  });
-
-  return { ...review, feedback_points: sanitisedLineNumbers };
-}
